% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRecommendation.R
\name{getRecommendation}
\alias{getRecommendation}
\title{Recommend movies}
\usage{
getRecommendation(jsonString)
}
\arguments{
\item{jsonString}{data passed on as \code{newdata} to \code{\link{predict}}}
}
\description{
Simple model to recommend movies
}
\examples{
mydata <- "[[34,5],[87, 2], [236, 3], [312, 4], [397, 1], [650, 2]]"
getRecommendation(mydata)
}

