% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recommender.R
\docType{data}
\name{recommender}
\alias{recommender}
\title{recommender model}
\format{a \code{\link{glm}} object}
\source{
Based on MovieLens data
}
\usage{
recommender
}
\description{
Simple model that recommends movies
}

