% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRecommendation.R
\name{getRecommendation}
\alias{getRecommendation}
\title{Recommend movies}
\usage{
getRecommendation(movies, ratings)
}
\arguments{
\item{movies}{data passed on to \code{\link{predict}}}

\item{ratings}{data passed on to \code{\link{predict}}}
}
\description{
Simple model to recommend movies
}
\examples{
moviesP <- "[34, 87, 236, 312, 397, 650]"
}

